DM lab1 shift hash implementation

+++ 6sem/DB/lab1/core/digit_analyze_hash_table.cpp
@@ -1,12 +1,12 @@
 #include "digit_analyze_hash_table.h"
 
-long DigitAnalyzeHashTable::calc_hash(long key)
+unsigned DigitAnalyzeHashTable::calc_hash(unsigned key)
 {
   // TODO: calcing digit analyze hash algorithm
   return 0;
 }
 
-int DigitAnalyzeHashTable::scale_hash(long hash)
+unsigned DigitAnalyzeHashTable::scale_hash(unsigned hash)
 {
   // TODO: calcing digit analyze scale algorithm
   return 0;
